requirements.txt
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
streamlit==1.9.0
scikit-learn>0.21.3
